i386, KVM: Add cpuid behavior for kvm xo
Add CPUID leaf FEAT_KVM_GENERIC for implementation of KVM based features
that can be implemented by other hypervisors.

Bit 1 of EAX on this leaf describes the KVM XO capability, which turns
the highest physical address bit supported into a virtual address
permission bit for the guest. In the presense of this XO bit, the
max physical address bits for the guest should be one less than the max
physical address of the host, so the top bit physical address bit can
become be used as a permission bit in the guest page tables.

Using this feature is only possible if KVM supports
KVM_CAP_EXECONLY_MEM, so print and error message if this is not the
case.

Signed-off-by: Rick Edgecombe <rick.p.edgecombe@intel.com>